Moved from compat addon to the actual addon
