+++ addons/armor/XtdGear/Uniforms_Ranks_Infantry.hpp
