+++ addons/armor/XtdGear/Uniforms_Tanker.hpp
