+++ addons/armor/XtdGearInfos.hpp
@@ -1,0 +1,30 @@
+class XtdGearInfos
+{
+    class CfgWeapons
+    {
+        #include "XtdGear\Helmets_P1.hpp"
+        #include "XtdGear\Helmets_P1_Geonosis.hpp"
+        #include "XtdGear\Helmets_P1_Pilot.hpp"
+        #include "XtdGear\Helmets_P1_Tanker.hpp"
+
+        #include "XtdGear\Helmets_P2.hpp"
+        #include "XtdGear\Helmets_P2_Insulated.hpp"
+        #include "XtdGear\Helmets_P2_Pilot.hpp"
+        #include "XtdGear\Helmets_P2_Tanker.hpp"
+
+        #include "XtdGear\Helmets_Airborne.hpp"
+        #include "XtdGear\Helmets_Engineer.hpp"
+
+        #include "XtdGear\Helmets_ARC.hpp"
+        #include "XtdGear\Helmets_ARF.hpp"
+        #include "XtdGear\Helmets_BARC.hpp"
+
+        #include "XtdGear\Uniforms_Ranks_Infantry.hpp"
+        #include "XtdGear\Uniforms_Ranks_Geonosis.hpp"
+        #include "XtdGear\Uniforms_P2_Insulated.hpp"
+        #include "XtdGear\Uniforms_Pilot.hpp"
+        #include "XtdGear\Uniforms_Tanker.hpp"
+        #include "XtdGear\Uniforms_ARC.hpp"
+        #include "XtdGear\Uniforms_ARF.hpp"
+    };
+};
