--- addons/compats/armor_aceax/XtdGearInfos.hpp
@@ -1,33 +1,0 @@
-class XtdGearInfos
-{
-    class CfgWeapons
-    {
-        #include "configs\Helmets_P1.hpp"
-        #include "configs\Helmets_P1_Geonosis.hpp"
-        #include "configs\Helmets_P1_Pilot.hpp"
-        #include "configs\Helmets_P1_Tanker.hpp"
-
-        #include "configs\Helmets_P2.hpp"
-        #include "configs\Helmets_P2_Insulated.hpp"
-        #include "configs\Helmets_P2_Pilot.hpp"
-        #include "configs\Helmets_P2_Tanker.hpp"
-
-        #include "configs\Helmets_Airborne.hpp"
-        #include "configs\Helmets_Engineer.hpp"
-
-        #include "configs\Helmets_ARC.hpp"
-        #include "configs\Helmets_ARF.hpp"
-        #include "configs\Helmets_BARC.hpp"
-
-        #include "configs\Uniforms_Ranks_Infantry.hpp"
-        #include "configs\Uniforms_Ranks_Geonosis.hpp"
-        #include "configs\Uniforms_P2_Insulated.hpp"
-        #include "configs\Uniforms_Pilot.hpp"
-        #include "configs\Uniforms_Tanker.hpp"
-        #include "configs\Uniforms_ARC.hpp"
-        #include "configs\Uniforms_ARF.hpp"
-
-        #include "configs\Uniforms_Custom.hpp"
-        #include "configs\Uniforms_Custom_Insulated.hpp"
-    };
-};
