--- addons/compats/armor_aceax/config.cpp
@@ -1,26 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches
-{
-    class SUBADDON
-    {
-        author = "Keeli Company Aux Team";
-        name = COMPONENT_NAME;
-        addonRootClass = QUOTE(ADDON);
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] =
-        {
-            QCLASS(armor),
-            QCLASS(custom_armor),
-            "aceax_main"
-        };
-        units[] = {};
-        weapons[] = {};
-        VERSION_CONFIG;
-
-        skipWhenMissingDependencies = TRUE;
-    };
-};
-
-#include "XtdGearModels.hpp"
-#include "XtdGearInfos.hpp"
