--- addons/compats/armor_aceax/script_component.hpp
@@ -1,4 +1,0 @@
-#define SUBCOMPONENT armor_aceax
-#define SUBCOMPONENT_BEAUTIFIED Armor - ACE Arsenal Extended
-
-#include "..\script_component.hpp"
