+++ addons/custom_armor/XtdGearInfos.hpp
@@ -1,0 +1,8 @@
+class XtdGearInfos
+{
+    class CfgWeapons
+    {
+        #include "XtdGear\Uniforms.hpp"
+        #include "XtdGear\Uniforms_P2_Insulated.hpp"
+    };
+};
