+++ addons/custom_armor/config.cpp
@@ -266,3 +266,5 @@ class CfgPatches
 #include "CfgWeapons.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgEventHandlers.hpp"
+#include "XtdGearModels.hpp"
+#include "XtdGearInfos.hpp"
